digi_log_be/requirements.txt
django
django-auth-ldap
django-cors-headers
djangorestframework
djangorestframework-simplejwt